refactor(history): extract shared InvoiceCard, remove duplicate card components
Unified HistoryEntryCard and ReceivedInvoiceCard (~95% identical) into a
single InvoiceCard with a nameLabel prop. Removed defeated memo(), pointless
useMemo, dead source prop, and unnecessary JSX comments. Collapsed identical
DecodedReceivedInvoice type into alias for DecodedHistoryEntry.
